+++ src/components/listView/ListView.jsx
@@ -1,6 +1,7 @@
 import styles from './list-view.module.css';
 import ListViewItem from '../listViewItem/ListViewItem';
 import { useState, useEffect } from 'react';
+import calculateDistance from '../../locationHelpers/calcDistance';
 
 function ListView({
   laundryServices,
@@ -44,25 +45,6 @@ function ListView({
       );
     }
   };
-
-  function calculateDistance(lat1, lon1, lat2, lon2) {
-    const R = 3959; // Earth's radius in miles
-
-    const dLat = ((lat2 - lat1) * Math.PI) / 180;
-    const dLon = ((lon2 - lon1) * Math.PI) / 180;
-
-    const a =
-      Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-      Math.cos((lat1 * Math.PI) / 180) *
-        Math.cos((lat2 * Math.PI) / 180) *
-        Math.sin(dLon / 2) *
-        Math.sin(dLon / 2);
-
-    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-    const distance = R * c; // Distance in miles
-
-    return distance;
-  }
 
   function extractSponsoredServices(laundryArray) {
     const sponsoredServicesArr = [];
